+++ app/Http/Controllers/Api/GroupController.php
@@ -7,11 +7,14 @@ use App\Http\Resources\GroupResource;
 use App\Models\Department;
 use App\Models\Group;
 use Auth;
+use Illuminate\Support\Facades\Gate;
 
 class GroupController extends BaseController
 {
     public function index()
     {
+        Gate::authorize('viewAny', Group::class);
+
         $groups = Group::all();
 
         $data = GroupResource::collection($groups);
@@ -19,8 +22,21 @@ class GroupController extends BaseController
         return $this->successResponse($data);
     }
 
+    public function indexByDepartment(Department $department)
+    {
+        Gate::authorize('viewByDepartment', [Group::class, $department]);
+
+        $data = GroupResource::collection(
+            Group::whereDepartmentId($department->id)->get()
+        );
+
+        return $this->successResponse($data);
+    }
+
     public function store(GroupRequest $request)
     {
+        Gate::authorize('create', Group::class);
+
         $validated = $request->validated();
 
         $group = Group::create($validated);
@@ -32,6 +48,8 @@ class GroupController extends BaseController
 
     public function show(Group $group)
     {
+        Gate::authorize('view', Group::class);
+
         $data = new GroupResource($group);
 
         return $this->successResponse($data);
@@ -39,6 +57,8 @@ class GroupController extends BaseController
 
     public function update(GroupRequest $request, Group $group)
     {
+        Gate::authorize('update', $group);
+
         $validated = $request->validated();
 
         $group->update($validated);
@@ -50,6 +70,8 @@ class GroupController extends BaseController
 
     public function destroy(Group $group)
     {
+        Gate::authorize('delete', $group);
+
         $group->delete();
 
         return $this->noContentResponse();
